fix: Clear deprecated constants on HA 2025.1
```
TIME_MINUTES was used from tuya_ble, this is a deprecated constant which will be removed in HA Core 2025.1. Use UnitOfTime.MINUTES instead, please report it to the author of the 'tuya_ble' custom integration
TIME_SECONDS was used from tuya_ble, this is a deprecated constant which will be removed in HA Core 2025.1. Use UnitOfTime.SECONDS instead, please report it to the author of the 'tuya_ble' custom integration
VOLUME_MILLILITERS was used from tuya_ble, this is a deprecated constant which will be removed in HA Core 2025.1. Use UnitOfVolume.MILLILITERS instead, please report it to the author of the 'tuya_ble' custom integration
TEMP_CELSIUS was used from tuya_ble, this is a deprecated constant which will be removed in HA Core 2025.1. Use UnitOfTemperature.CELSIUS instead, please report it to the author of the 'tuya_ble' custom integration
```